move Colleciton to correct namespace
